libkvs/kvs_txn: Refactor kvs_txn
Splice flux_kvs_fence() and flux_kvs_commit() out from kvs_txn
and into new kvs_commit files.
