+++ src/common/libkvs/kvs_commit.h
@@ -1,0 +1,17 @@
+#ifndef _FLUX_CORE_KVS_COMMIT_H
+#define _FLUX_CORE_KVS_COMMIT_H
+
+enum kvs_commit_flags {
+    FLUX_KVS_NO_MERGE = 1, /* disallow commits to be mergeable with others */
+};
+
+flux_future_t *flux_kvs_commit (flux_t *h, int flags, flux_kvs_txn_t *txn);
+
+flux_future_t *flux_kvs_fence (flux_t *h, int flags, const char *name,
+                               int nprocs, flux_kvs_txn_t *txn);
+
+#endif /* !_FLUX_CORE_KVS_COMMIT_H */
+
+/*
+ * vi:tabstop=4 shiftwidth=4 expandtab
+ */
