+++ src/common/libkvs/kvs_txn_private.h
@@ -1,8 +1,11 @@
 #ifndef _KVS_TXN_PRIVATE_H
 #define _KVS_TXN_PRIVATE_H
 
-/* interface for unit tests only */
-enum { TXN_GET_FIRST, TXN_GET_NEXT, TXN_GET_ALL };
+enum {
+    TXN_GET_FIRST,
+    TXN_GET_NEXT,
+    TXN_GET_ALL
+};
 int txn_get (flux_kvs_txn_t *txn, int request, void *arg);
 
 #endif /* !_KVS_TXN_PRIVATE_H */
